feat: rename folder and files for back-end section
